Another SQL DataStore fix to delete simulations in batches from .db tables

+++ Models/DataStore.cs
@@ -261,22 +261,32 @@ namespace Models
                 foreach (string simulationNameToBeRun in simulationNamesToBeRun)
                     idsToDelete.Add(GetSimulationID(simulationNameToBeRun));
 
-                idString = "";
+                idString = string.Empty;
+                j = 0;
                 for (int i = 0; i < idsToDelete.Count; i++)
                 {
-                    if (i > 0)
+                    if (j > 0)
                         idString += " OR ";
                     idString += "SimulationID = " + idsToDelete[i].ToString();
+
+                    if (j == 100 || j == idsToDelete.Count-1)
+                    {
+                        foreach (string tableName in TableNames)
+                        {
+                            // delete this simulation
+                            RunQueryWithNoReturnData("DELETE FROM " + tableName + " WHERE " + idString);
+                        }
+
+                        if (TableNames.Contains(UnitsTableName))
+                            RunQueryWithNoReturnData("DELETE FROM " + UnitsTableName + " WHERE " + idString);
+
+                        idString = string.Empty;
+                        j = 0;
+                    }
+                    else
+                        j++;
                 }
 
-                foreach (string tableName in TableNames)
-                {
-                    // delete this simulation
-                    RunQueryWithNoReturnData("DELETE FROM " + tableName + " WHERE " + idString);
-                }
-
-                if (TableNames.Contains(UnitsTableName))
-                    RunQueryWithNoReturnData("DELETE FROM " + UnitsTableName + " WHERE " + idString);
             }
             finally
             {
